feat(firmware)!: let a project raise the I/O buffer sizes its board compiles
DOPE-370. A P1AM-200 with 256 KB of SRAM served the same 56 coils as an Uno
with 2 KB, because MAX_DIGITAL_OUTPUT was a literal in openplc.h chosen per MCU
family. Every one of those macros is now `#ifndef`-guarded, and openplc.h picks
up a generated io_sizes.h when the project asked for more room.

io_sizes.h has to reach the build through openplc.h rather than defines.h.
openplc.h is included by ~20 HAL translation units that never see defines.h,
and defines.h deliberately carries no include guard so it can only arrive
through exactly one path (modbus_config.h explains why). A separately guarded
file is the only way to reach every consumer without breaking that rule; it is
pulled in behind `__has_include` so its absence is not an error.

The register store was capped at 255 in three separate places, all uint8_t:
init_mbregs's six parameters, MBinfo's six *_size fields, and the byte_addr /
pos indices into the packed bit banks. Asking for 256 coils wrapped to 0 and
every bounds check then rejected every address -- the shape of the P1AM report.
All are uint16_t now, which is also the natural ceiling: a Modbus address is 16
bits, so a count that did not fit could not be reached on the wire.

Two rules keep this safe:

  * Growth only. These macros dimension the IEC pointer arrays as well as the
    Modbus banks, and mapEmptyBuffers() aliases %MW/%MD/%ML into those banks,
    so a segment shrunk below what a compiled program addresses would drop I/O
    with no diagnostic. The floor is the board's firmware default.
  * Silence by default. When every count equals the board default,
    generateIoSizesHeader returns null, no file is written, and the firmware
    compiles byte-for-byte as it did.

The ceiling comes from the board's package (`device.ioMax`). A board that
declares none cannot be raised at all, and the screen renders its counts
read-only rather than offering an input that cannot move.

+++ resources/sources/Baremetal/modbus_registers.cpp
@@ -9,7 +9,7 @@ Copyright (C) 2022 OpenPLC - Thiago Alves
 // In a debug-only build this whole TU compiles to nothing, saving flash/SRAM.
 #ifdef MODBUS_ENABLED
 
-bool init_mbregs(uint8_t size_holding, uint8_t size_dint_memory, uint8_t size_lint_memory, uint8_t size_coils, uint8_t size_inputregs, uint8_t size_inputstatus)
+bool init_mbregs(uint16_t size_holding, uint16_t size_dint_memory, uint16_t size_lint_memory, uint16_t size_coils, uint16_t size_inputregs, uint16_t size_inputstatus)
 {
     //Save sizes
     modbus.holding_size = size_holding;
@@ -64,7 +64,9 @@ bool init_mbregs(uint8_t size_holding, uint8_t size_dint_memory, uint8_t size_li
 
 bool get_discrete(uint16_t addr, bool regtype)
 {
-    uint8_t byte_addr = addr / 8;
+    // byte_addr indexes the packed bit bank, so it is addr/8 of a 16-bit
+    // address -- it does not fit in a uint8_t past 2040 bits.
+    uint16_t byte_addr = addr / 8;
     uint8_t bit_addr = addr % 8;
     if (regtype == COILS)
         return bitRead(modbus.coils[byte_addr], bit_addr);
@@ -74,7 +76,7 @@ bool get_discrete(uint16_t addr, bool regtype)
 
 void write_discrete(uint16_t addr, bool regtype, bool value)
 {
-    uint8_t byte_addr = addr / 8;
+    uint16_t byte_addr = addr / 8;
     uint8_t bit_addr = addr % 8;
     if (regtype == COILS)
         bitWrite(modbus.coils[byte_addr], bit_addr, value);
@@ -116,7 +118,7 @@ void readRegisters(uint16_t startreg, uint16_t numregs)
 
     uint16_t val;
     uint16_t i = 0;
-    uint8_t pos = 0;
+    uint16_t pos = 0;
 	while(numregs--)
     {
         if ((startreg + i) < modbus.holding_size)
@@ -177,7 +179,7 @@ void writeSingleRegister(uint16_t reg, uint16_t value)
         return;
     }
 
-    uint8_t pos = 0;
+    uint16_t pos = 0;
 
     if (reg < modbus.holding_size)
     {
@@ -254,7 +256,7 @@ void writeMultipleRegisters(uint16_t startreg, uint16_t numoutputs, uint8_t byte
 
     uint16_t value;
     uint16_t i = 0;
-    uint8_t pos = 0;
+    uint16_t pos = 0;
 	while(numoutputs--)
     {
         value = (uint16_t)mb_frame[7+i*2] << 8 | (uint16_t)mb_frame[8+i*2];
